:recycle: :: refactor components directory structures
