+++ src/app/components/thumbnail/large/large.module.css
