+++ src/app/components/thumbnail/normal/normal.module.css
