+++ src/app/components/thumbnail/row/row.module.css
